refactor: 검색어 도메인 엔티티 통합
