+++ src/main/java/com/recipe/app/src/recipe/domain/SearchKeyword.java
@@ -1,20 +1,18 @@
-package com.recipe.app.src.recipe.infra.keyword;
+package com.recipe.app.src.recipe.domain;
 
 import lombok.AccessLevel;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
-@NoArgsConstructor(access = AccessLevel.PUBLIC)
-@EqualsAndHashCode(callSuper = false)
-@Data
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
 @Table(name = "Keywords")
-public class SearchKeywordEntity {
+public class SearchKeyword {
+
     @Id
     @Column(name = "keyword", nullable = false, updatable = false)
     private String keyword;
-
 }
