--- src/main/java/com/recipe/app/src/recipe/infra/keyword/SearchKeywordJpaRepository.java
@@ -1,6 +1,0 @@
-package com.recipe.app.src.recipe.infra.keyword;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface SearchKeywordJpaRepository extends CrudRepository<SearchKeywordEntity, String> {
-}
