+++ src/main/java/com/recipe/app/src/recipe/infra/keyword/SearchKeywordRepository.java
@@ -1,0 +1,9 @@
+package com.recipe.app.src.recipe.infra.keyword;
+
+import com.recipe.app.src.recipe.domain.SearchKeyword;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SearchKeywordRepository extends JpaRepository<SearchKeyword, String> {
+}
